moved pages dir
