+++ pages/api/assets.ts
